:white_check_mark: test: 테스트 코드 분리
